rename oppsummering av behandling og vedtak
Dem heter nå Søknadsbehandlingoppsummering, og søknadsbehandlingvedtakoppsummering

Dette er for å bedre skille hvilken del av saksoversikt flyten dem er i
