+++ src/components/søknadsbehandlingoppsummering/søknadsbehandling-nb.ts
