+++ src/components/søknadsbehandlingoppsummering/søknadsbehandlingHeader.module.less
